Agregando el servicio del carrito, la implementacion del servicio y su controlador

+++ src/main/java/com/spring/eccomerce/dto/producto/ProductoFiltroDTO.java
@@ -4,13 +4,15 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Builder
 @Setter @Getter
 public class ProductoFiltroDTO {
 
     private Long categoriaId;
     private String nombre;
-    private Double precioMinimo;
-    private Double precioMaximo;
+    private BigDecimal precioMinimo;
+    private BigDecimal precioMaximo;
     private Integer existencia = 0;
 }
